rename standalone_server to standalone_storage
