+++ kv/storage/standalone_storage/standalone_storage.go
@@ -1,4 +1,4 @@
-package standalone_server
+package standalone_storage
 
 import (
 	"github.com/Connor1996/badger"
